Switched back to 32bit nodes

+++ include/defs.h
@@ -41,7 +41,7 @@ using uint_t = std::uint64_t;
 using external_size_t = uint_t;
 
 //using node_t = int_t; ///< Type for every node id used in this project
-using node_t = int64_t;
+using node_t = int32_t;
 constexpr node_t INVALID_NODE = std::numeric_limits<node_t>::max();
 
 using degree_t = int32_t; ///< Type for node degrees
